Use multi service object to get data type
When reading/writing tags, we have to get the data type first, mainly
because of boolean arrays and reading bits of a word.  This only happens
once per uniquee tag, but if you are reading a lot of unique tags, this
takes time.  This change speeds things up because it bunldles multiple
tags into one packet.