Fixed reading struct larger than one packet
There were several problems here with how partial transfers were
being handled.  I restructured how the packets were collected and parsed.
Originally, we would send a request then parse the reply, if there was more
data coming, update the offset, then resend the request to get the next
chunk.  This was tricky to manage when it came to UDTs.

Now we gather all of the raw data that span over multiple packets then
parse it after all of the data is gathered.  The process is simpler
and better handles the different scenarios (single tag, array,
large array, struct, etc).

[Issue #142]