Fixed providing data type when reading lists
Providing the data type when reading lists was broken when I
switched to utilizing multi-read service to discover the data type.
I was passing the tuple/list to the initial read as the tag name. I also
needed to save the tag and data type in KnownTags when the data type
is provided so that we don't try to retreive the data type when it
was provided to us.