Fixed issue array BOOL's in a list
When reading BOOL arrays in a list, the array index was being
utilized for the member.  This should have been the array
index/32 because BOOL arrays are not stored in the PLC in the
way that they are displayed to the user.  They are stored in
4 byte chunks. Behind the scenes, we have to treat them similar
to accessing bits of a DINT array.  For example if we want to
read [63] of a BOOL array, we don't request member 63, we have
to request member 1, then mask out bit 31 from the value the PLC
returns and send that to the user.