Fixed crashes when reading/writing in a loop an connection is lost
After a successful read, if the PLC is disconnected, there was no check
to be sure that good data was returned from the controller, which
would cause a crash.  Fixed single reads/writes as well as multi
service requests.  Connection status is returned in the status.

Also changed, when reading lists and we crashed in the way described above,
a single instance of response was returned.  Now an instance per tag is
returned.