Fixed crash if trying to wite to a tag that doesn't exist
If you were to write to a tag that was STRING type that didn't exist,
instead of getting a Path Segment Error, pylogix would crash.  This is
because if the datatype is unknown, it assumed that the value was an
integer, when it is actually a string, in this case.  So rather than
assume int, we look the value type.