Write arrays when writing lists of tags/values
When writing lists of tags, if an item was attempting to
write a list of values, only the first value would be written.
This is because only the first value was passed to the method
generating the request.  This will pass all of the values, allowing
writing the whole list of values

[Issue #136]