Fixes for reading large lists
Reading large lists, ones that are larger than a single packet, weren't properly
predicting the size of the reply.  So it was possible to send more data in the request
than would fit in the reply.  While most data types worked fine, STRING's did not.
This is likely because STRING's return the entire 82 bytes regardless of how long
the actual STRING data actually is.  We need to consider the reply when sending the
request.  So the length of the tag name matters and the data type matters.  The reply will
always be larger than the request because the request contains the value, so calculate
what the reply will be to know when we need to start a new multi-service request.