Fix getting module properties
A recent change returned only slot 0's module properties, no matter
which slot was requested.  The problem has to to with the solution to request
properties for devices that are not PLC's.  If not route is supplied, than
the fact that a user provided a slot is ignored.  This change considers whether
the user actually provided a slot number by defaulting slot to None