Fixes for writing large BOOL arrays
My nemesis.  The main problem here is that BOOL arrays are packed in
4 byte chunks.  While you see an array of bools in RSLogix, internally,
they are stored more like an array of DINT's, 32 BOOL's per DINT.  That
is why you are forced to create BOOL arrays in 4 byte increments.

When you write multiple BOOLS, there are a few things to consider:
  - What is the starting BOOL (MyTag[4])
  - How many BOOL's are being written
  - Do they all fit in the current 4 bytes

The solution here is to generate multiple 4 byte writes.  First we have
to find the boundaries (start bool, number of bools to write), then
generate the appropriate number of writes to handle them all.  The modified
write is only able to handle 4 bytes at a time, so if the user tries to
write to 2 or more words, we have to generate 2 or more writes.

The key piece was that the function that generates the IOI also packs the
index ([n])  Once that happened, we were committed to the first word that
we were attempting to write to.  The solution to this was to internally
generate a new tag list to re-generate IOI, in order to get to the next
word.  For example, lets say we're writing 64 BOOLs to MyTag[0]. We would
need to writes.  Internally, we generate a write for MyTag[0] to write the
first 32 bits, then MyTag[32] for the next 32 bits.  The user only provides
writing to MyTag[0].

[Issue: 190]