Fixes for writing bool arrays
I was calculating how much data could fit in a packet incorrectly
for boolean array writes.  This is because when writing data into
BOOL arrays, data gets packed into 4 byte chunks.  The result was
that when large enough writes were requested, fragmented write service
was used because we calculated that the data wouldn't fit in the packet.

For example, writing 32 bool will take at most 8 bytes, if it spans
2 words.  I was calculating that it would take 32 * 4 bytes in this
case.  If you chose enough BOOLS, it would trigger the fragmented write.

The reality is that when writing BOOL arrays, they are broken up into
writes that deal with 4 bytes at a time, so fragmented write should
never be used.

[Issue #190]