Fixes for writing large lists
Large lists didn't span multiple packets.  This change adds "batch write"
similar to the read function, which estimates what fits in the packet
and generates multiple multi-service writes.

BOOL arrays were of course the sticking point.  Not only to we have to
predict if they will fit into the packet, but as we are assembling the
requests, if we run out of room, we have put the request in the next
packet.