Fix for reading lists.
Micro800 wasn't handled correctly.  The list, whiich included the tag name, length
and data type was passed as the tag name, so that list was passed to discover
unknown tags, rather than just the tag name.  Also, for Logix controllers, the
discovering unknown tags was looking at the length value rather than data type.  So
when you provided tag_name, length, data type, it would use the length as the data type.
Consequently, if you just provided tag_name and data_type, it would "work". This was
not the way it was intended.