Updates to how ioi is generated
There was a condition for multi-dimension arrays that wasn't handled
correctly.  The specific example was reading the LEN member of a
multi-dimension string.  For a tag like Multi[1,1,1], we could simply
read "Multi" to get the data type.  However, with a multi-dimension string,
if you try to read the LEN: Multi[1,1,1].LEN, when trying to get the data type
we ingored the extra dimensions, which caused a path segment error.  This change
addresses the issue and cleans up the function.